Added augustus/julius

+++ augustus/macsourceports_universal2.sh
@@ -27,24 +27,40 @@ git checkout tags/${GIT_TAG}
 
 rm -rf ${BUILT_PRODUCTS_DIR}
 
-# create makefiles with cmake, perform builds with make
-rm -rf ${X86_64_BUILD_FOLDER}
-mkdir ${X86_64_BUILD_FOLDER}
-cd ${X86_64_BUILD_FOLDER}
-/usr/local/bin/cmake ..
-make -j$NCPU
+if [ "$1" == "buildserver" ] || [ "$2" == "buildserver" ]; then
+    # create makefiles with cmake, perform builds with make
+    mkdir ${BUILT_PRODUCTS_DIR}
+    cd ${BUILT_PRODUCTS_DIR}
+    cmake "-DCMAKE_OSX_ARCHITECTURES=arm64;x86_64" ..
+    cmake --build . --parallel $NCPU
+    install_name_tool -add_rpath @executable_path/. ${EXECUTABLE_FOLDER_PATH}/${EXECUTABLE_NAME}
+    cp /usr/local/lib/libSDL2-2.0.0.dylib ${EXECUTABLE_FOLDER_PATH}
+    cp /usr/local/lib/libSDL2_mixer-2.0.801.0.0.dylib ${EXECUTABLE_FOLDER_PATH}
+    cp /usr/local/lib/libpng16.16.dylib ${EXECUTABLE_FOLDER_PATH}
+else
+    # create makefiles with cmake, perform builds with make
+    rm -rf ${X86_64_BUILD_FOLDER}
+    mkdir ${X86_64_BUILD_FOLDER}
+    cd ${X86_64_BUILD_FOLDER}
+    /usr/local/bin/cmake ..
+    make -j$NCPU
 
-cd ..
-rm -rf ${ARM64_BUILD_FOLDER}
-mkdir ${ARM64_BUILD_FOLDER}
-cd ${ARM64_BUILD_FOLDER}
-cmake ..
-make -j$NCPU
+    cd ..
+    rm -rf ${ARM64_BUILD_FOLDER}
+    mkdir ${ARM64_BUILD_FOLDER}
+    cd ${ARM64_BUILD_FOLDER}
+    cmake ..
+    make -j$NCPU
+fi
 
 cd ..
 
 # create the app bundle
-"../MSPBuildSystem/common/build_app_bundle.sh"
+if [ "$1" == "buildserver" ] || [ "$2" == "buildserver" ]; then
+    "../MSPBuildSystem/common/build_app_bundle.sh" "skiplipo"
+else
+    "../MSPBuildSystem/common/build_app_bundle.sh"
+fi
 
 #sign and notarize
 "../MSPBuildSystem/common/sign_and_notarize.sh" "$1"
